Move get_current_ip function from app.py to ipget module
